updated utils
