+++ src/parsing/utils.c
@@ -1,109 +1,5 @@
 #include "minirt.h"
 
-static int	ft_atoi_and_overflow(const char *nptr, int *overflow);
-
-int	parse_error(t_scene *scene, char *msg, char **res, t_shape *shape)
-{
-	if (msg)
-		error_msg(msg, scene->error_line);
-	if (res)
-		free_array(res);
-	if (shape)
-		free(shape);
-	return (1);
-}
-int	parse_fatal(t_scene *scene, int fd)
-{
-	if (fd >= 0)
-		close(fd);
-	free_scene(scene);
-	return (1);
-}
-
-int	error_return(char *msg, char *line)
-{
-	error_msg(msg, line);
-	return (1);
-}
-void	error_msg(char *msg, char *line)
-{
-	ft_putendl_fd(ERR_MSG, 2);
-	ft_putendl_fd(msg, 2);
-	if (line != NULL)
-		ft_printf("Problem in line: %s\n", line);
-}
-
-void	free_scene(t_scene *scene)
-{
-	free_list(scene->obj_list);
-	scene->obj_list = NULL;
-}
-
-void	free_list(t_olist *list)
-{
-	t_olist	*temp;
-
-	if (!list)
-		return ;
-	while (list)
-	{
-		temp = list->next;
-		if (list->shape)
-		{
-			free_matrix(list->shape->matrix);
-			free(list->shape);
-		}
-		free(list);
-		list = temp;
-	}
-}
-
-int	add_to_list(t_scene *scene, t_shape *shape)
-{
-	t_olist	*new;
-	t_olist	*temp;
-
-	new = malloc(sizeof(t_olist));
-	if (!new)
-		return (1);
-	shape->obj_id = scene->next_obj_id++;
-	new->shape = shape;
-	new->next = NULL;
-	if (scene->obj_list == NULL)
-		scene->obj_list = new;
-	else
-	{
-		temp = scene->obj_list;
-		while (temp->next)
-			temp = temp->next;
-		temp->next = new;
-	}
-	return (0);
-}
-
-t_vec3	creat_vec3(float x, float y, float z)
-{
-	t_vec3	vector;
-
-	vector.x = x;
-	vector.y = y;
-	vector.z = z;
-	return (vector);
-}
-
-void	free_array(char **arr)
-{
-	int	i;
-
-	i = 0;
-	if (!arr)
-		return ;
-	while (arr[i])
-		free(arr[i++]);
-	free(arr);
-}
-
-// checks how many elements in the string
 int	count_elements(char *line)
 {
 	int	res;
@@ -207,175 +103,3 @@ int	is_valid_comas(const char *line)
 	return (0);
 }
 
-// with bitshifting storing rgb into singl int
-// t - transparency (always to max value = 255)
-int	ft_rgbtoint(int transp, int red, int green, int blue)
-{
-	return (transp << 24 | red << 16 | green << 8 | blue);
-}
-
-// ascii to float converter
-// detects sign
-// if overflow happens, still writes garbage value to res
-// returns 0 on success, 1 otherwise
-// might need to make res = 0;
-int	ft_atof(const char *line, float *res)
-{
-	int		sign;
-	int		i;
-	float	temp;
-	int		overflow;
-
-	i = 0;
-	sign = 1;
-	temp = 0.0f;
-	overflow = 0;
-	if (line[i] == '-' || line[i] == '+')
-	{
-		if (line[i] == '-')
-			sign *= -1;
-		i++;
-	}
-	get_whole_part(line, &i, &temp, &overflow);
-    if (!overflow)
-    {
-        get_fraction_part(line, &i, &temp);
-        *res = temp * sign;
-    }
-	return (overflow);
-}
-// part of ft_atof function.
-// gets the whole part of float num (before '.')
-// modifies passed params
-void	get_whole_part(const char *line, int *i, float *temp, int *overflow)
-{
-	
-	while (ft_isdigit(line[*i]))
-	{
-		*temp = *temp * 10 + (line[*i] - '0');
-		if (*temp > FLT_MAX) // for now, need to define by myself mb
-		    *overflow = 1;
-		(*i)++;
-	}
-}
-// part of ft_atof function.
-// gets the fractional part of float num (after '.')
-// modifies passed params
-void	get_fraction_part(const char *line, int *i, float *temp)
-{
-	float	fraction;
-
-	fraction = 0.1f;
-	if (line[*i] == '.')
-		(*i)++;
-	while (ft_isdigit(line[*i]))
-	{
-		*temp += (line[*i] - '0') * fraction;
-		fraction *= 0.1f;
-		(*i)++;
-	}
-}
-
-int	parse_rgb(char *str, int *color)
-{
-	char	**rgb;
-	int		r;
-	int		g;
-	int		b;
-
-	if (!str || is_valid_comas(str))
-		return (1);
-	rgb = ft_split(str, ',');
-	if (!rgb)
-		return (error_return(ERR_ALLOC, NULL));
-	if (!rgb[0] || !rgb[1] || !rgb[2] || rgb[3])
-		return (free_array(rgb), 1);
-	r = 0;
-	g = 0;
-	b = 0;
-	if (parse_int(rgb[0], 0, 255, &r) == 1 || parse_int(rgb[1], 0, 255, &g) == 1
-		|| parse_int(rgb[2], 0, 255, &b) == 1)
-		return (free_array(rgb), 1);
-	*color = ft_rgbtoint(255, r, g, b);
-	free_array(rgb);
-	return (0);
-}
-
-int	parse_vector(char *str, t_vec3 *vector, float min, float max)
-{
-	char	**res;
-	float	x;
-	float	y;
-	float	z;
-
-	if (!str || is_valid_comas(str) == 1)
-		return (1);
-	res = ft_split(str, ',');
-	if (!res)
-		return (error_return(ERR_ALLOC, NULL));
-	if (!res[0] || !res[1] || !res[2] || res[3])
-		return (free_array(res), 1);
-	x = 0;
-	y = 0;
-	z = 0;
-	if (parse_float(res[0], min, max, &x) == 1 || parse_float(res[1], min, max,
-			&y) == 1 || parse_float(res[2], min, max, &z) == 1)
-		return (free_array(res), 1);
-	*vector = creat_vec3(x, y, z);
-	free_array(res);
-	return (0);
-}
-
-int	parse_int(char *str, int min, int max, int *res)
-{
-	int	temp;
-	int	overflow;
-
-	overflow = 0;
-	temp = 0;
-	if (!str || is_valid_int(str) == 1)
-		return (1);
-	temp = ft_atoi_and_overflow(str, &overflow);
-	if (temp < min || temp > max || overflow == 1)
-		return (1);
-	*res = temp;
-	return (0);
-}
-
-int	parse_float(char *str, float min, float max, float *res)
-{
-	if (!str || is_valid_float(str) == 1)
-		return (1);
-	if (ft_atof(str, res))
-		return (1);
-	if (*res < min || *res > max)
-		return (1);
-	return (0);
-}
-static int	ft_atoi_and_overflow(const char *nptr, int *overflow)
-{
-	int		i;
-	long	num;
-	long	neg;
-
-	i = 0;
-	num = 0;
-	neg = 1;
-	while ((nptr[i] >= 9 && nptr[i] <= 13) || nptr[i] == 32)
-		i++;
-	if (nptr[i] == '-' || nptr[i] == '+')
-	{
-		if (nptr[i] == '-')
-			neg *= -1;
-		i++;
-	}
-	while (nptr[i] >= '0' && nptr[i] <= '9')
-	{
-		num = num * 10 + (nptr[i] - '0');
-		if ((num > INT_MAX && neg == 1) || (num > (long)INT_MAX + 1 && neg ==
-				-1))
-			*overflow = 1;
-		i++;
-	}
-	return ((int)num * neg);
-}
